Restore now installs dependencies
The nuget.exe restore command doesn't install dependencies, it only
installs what is in the packages.config.  nuget.exe doesn't auto-manage
the packages.config, so we must do so.  It is easier to simply force the
Install method to be called for each package, since Install also
installs dependencies.

The packages.config file will need to be updated with dependencies soon,
since that is what the GUI uses to know what is installed.