Dependencies are added to the packages.config file
When packages with dependencies are installed, the dependencies are also
added to the pacakges.config file.  This is done through a parsing of
the output text to see if nuget.exe installed dependencies.  It feels
very hacky, but it is what must be done since there  is no API for
manipulating the .config file or getting dependencies.