The search field now searches instead of filters
The search field used to filter the results because a list of ALL
packages used to be retrieved from the server.

This is because that is all nuget.exe allowed for. Since the HTTP method
of searching allows specifying how many results to get, we can do a full
"real" search of the server.