Various fixes
The first version returned by the server (usually the latest version) is
now used if a specific version of a package isn't found.

When deleting directories (during the Clean operation) read-only files
are now automatically marked writable to prevent exceptions and allow
the directory to be properly deleted.

Compact Framework, Silverlight, and Windows Phone DLLs will no longer be
imported, if they exist in a package.

.NET 3.0 DLLs will no longer be imported if a .NET 3.5 DLL is found in
the package.

Added a (currently unused) Verbose flag for outputting debug
information.